Nibba
Prostate cancer is the development of cancer in the prostate, a gland in the male reproductive system.[6] Most prostate cancers are slow growing; however, some grow relatively quickly. The cancer cells may spread from the prostate to other area of the body, particularly the bones and lymph nodes. It may initially cause no symptoms. In later stages it can lead to difficulty urinating, blood in the urine, or pain in the pelvis, back or when urinating. A disease known as benign prostatic hyperplasia may produce similar symptoms. Other late symptoms may include feeling tired due to low levels of red blood cells.